address comments pt. 2
