+++ micro-rdk/src/native/mod.rs
@@ -1,6 +1,7 @@
 pub mod certificate;
 pub mod dtls;
 pub mod entry;
+pub mod log;
 pub mod tcp;
 pub mod tls;
 pub mod conn {
